requirements.txt
streamlit
openai
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
groq